logic separation from amity.py to app.py and run.py
