+++ .travis.yml
@@ -1,0 +1,12 @@
+language: python
+python:
+  - "2.7"
+install:
+  - pip install -r requirements.txt
+  - pip install coveralls
+env:
+  - DB=sqlite
+script: nosetests --with-coverage
+
+after_success:
+ - coveralls
