+++ requirements.txt
@@ -1,0 +1,22 @@
+args==0.1.0
+backports.shutil-get-terminal-size==1.0.0
+clint==0.5.1
+colorama==0.3.7
+coverage==4.0.3
+coveralls==1.1
+decorator==4.0.9
+docopt==0.6.2
+ipdb==0.9.3
+ipython==4.2.0
+ipython-genutils==0.1.0
+nose==1.3.7
+pathlib2==2.1.0
+pexpect==4.0.1
+pickleshare==0.7.2
+ptyprocess==0.5.1
+pyfiglet==0.7.5
+requests==2.10.0
+simplegeneric==0.8.1
+six==1.10.0
+termcolor==1.1.0
+traitlets==4.2.1
